Finished a basic version of guild snipes??????
Not 100% sure if this actually works, should probably write some tests.
Also it is currently very barebones.
Also also, need to add a thingy to check if the guild snipes feature is enabled for the current guild - currently it's just enabled for everything. This would also require figuring out database migration tho.